MODIFY : pass S only test.
Former-commit-id: d5324360f41e81bc12a6f0c624491e18315fadd2

+++ project5/db_project/db/include/trx.h
@@ -19,7 +19,7 @@
 extern pthread_mutex_t trx_manager_latch;
 
 // Global Transaction ID 
-extern int global_trx_id;
+extern int64_t global_trx_id;
 
 // Transaction Manager
 class TrxManager {
